ADX file format support
Go's encoding/xml encodes all newlines as &#xA; even if they would be
valid in a chardata, so the test literal is uglier than necessary.
There also doesn't seem to be a way to have each element start on a
newline but not indented (Indent("\n", "") results in a blank line
between each element), so the text output doesn't match the test QSOs
adx file.