Support user-defined fields in ADI and ADX files
The CSV format doesn't have a good way to include userdef metadata and
I wasn't convinced that the natural way to represent them in the JSON
format used so far would be great either, so userdef parsing and output
are limited to ADI and ADX.  The --userdef option can define an
ephemeral field for validation for the other formats.

This change also cleans up some chunks of code encountered along the way.