Conditions and comparisons; add find and sort commands
This evolves `edit --if` to allow boolean and, or, and not combinations.
It uses the same if, if-not, or-if, or-if-not options and syntax for
the new find command, which filters a log.  The new sort command uses
the same comparisons to put records in order by an arbitrary list of
fields, with a - prefix indicating descending order.

Condition flags must have a field name on the left side, then an
operator, and then one or more values.  Multiple values are separated
by | characters.  Fields can be compared with other fields with
{FIELD_NAME} on the right-hand side.  Only =, <, <=, >, >= operators are
currently supported.  "Not equals" is currently expressed as
`--if-not field=value` rather than adding a not-equals operator.
I've thought some about flexible string match operators like being able
to match W1AW/P and W1AW, but I'm not sure what syntax to use.
Parsing SQL is a non-goal, but it's possible there's a more elegant
approach than using flags to express boolean logic here.