Improve primary subdivision (state/province) validation
Previously, if the ADIF spec didn't define a list of primary
administrative subdivisions for a country, adifmt validate would report
a warning if the STATE field value matched the code for some other
country and report an error if the value wasn't part of any country's
code list in the ADIF spec.  With this change, any code for a country
with no subdivision enum values will be a warning; for a country with a
non-empty subdivision enum, any value not matching the list will be an
error.  If the relevant DXCC or MY_DXCC field is not set, any value in
STATE/MY_STATE will result in a warning.