flatten: accept whitespace as a --delimiter value
This previously used the same FieldAssignments flag type for delimiter
that's used for edit --set and friends, which allows multiple values in
a single use like 'DXCC=291 ;; STATE=CT' (using ;; as an assignment
delimiter) and then trimming whitespace in the individual assignments.
This isn't a great fit for specifying delimiters for fields, since `;;`
itself could be a useful delimiter, and trimming space on the assignment
prevents the very sensible use of space as a delimiter.

With this change, each delimiter assignment must be specified as a
separate flag invocation: `adifmt flatten --delimiter 'foo=,'
--delimiter 'bar=|' --fields foo,bar`.